refactor: util 폴더 위치 변경에 따른 의존성 변경 [#169]
