+++ server/src/Email/Email.router.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { authValidator } from '../../lib/utils/authValidator';
+import { authValidator } from '../utils/authValidator';
 import { inviteByEmail, isValidEmail } from './Email.controller';
 
 const router = express.Router();
