+++ server/src/Tasks/Tasks.controller.ts
@@ -2,7 +2,7 @@ import { Request, Response } from 'express';
 import { getRepository } from 'typeorm';
 import Tasks from './Tasks.entity';
 import Stories from '../Stories/Stories.entity';
-import { bodyValidator } from '../../lib/utils/requestValidator';
+import { bodyValidator } from '../utils/requestValidator';
 
 export const getAllTasksByProject = async (req: Request, res: Response) => {
   try {
